search-backend/requirements.txt
pyyaml~=6.0
fastapi~=0.85.1
uvicorn[standard]
matplotlib~=3.6.1
pandas~=1.5.0
starlette~=0.20.4